+++ R/ms_grouped_y2.R
@@ -9,7 +9,7 @@
 #' @param group_var DEFAULT = 'group_var'; All levels of the group_var must be present or the chart may break. To do this, save the variable as_factor() before running freqs. Also remember that label_text and label_color must exactly match all the levels of the group_var or the function will break.
 #' @param axis_text_size DEFAULT = 14; Font size for variable levels and percentages.
 #' @param axis_title_size DEFAULT = 18; Font size for axis_x_label and axis_y_label.
-#' @param axis_x_display,axis_y_display DEFAULT = T
+#' @param axis_x_display,axis_y_display DEFAULT = TRUE
 #' @param axis_x_label,axis_y_label DEFAULT = ''; Title for the x_axis and y_axis
 #' @param axis_y_min DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.
 #' @param axis_y_max DEFAULT = NULL
@@ -22,7 +22,7 @@
 #' @param grouping DEFAULT = 'standard'; grouping for a barchart, a linechart or an area chart. must be one of "percentStacked", "clustered", "standard" or "stacked".
 #' @param label_color DEFAULT = color_settings_grouped; A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings_grouped <- list('Name of Group 1' = lime,'Name of Group 2' = brightblue)
 #' @param label_position DEFAULT = 'outEnd'; Other options include c('outEnd', 'inEnd', 'ctr', 'inBase')
-#' @param label_show_values DEFAULT = T; TRUE or FALSE. Show percent labels for each value.
+#' @param label_show_values DEFAULT = TRUE; TRUE or FALSE. Show percent labels for each value.
 #' @param label_text DEFAULT = text_settings_grouped; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings_grouped <- list('Name of Group 1' = fp_text(font.size = 16, color = lime),'Name of Group 2' = fp_text(font.size = 16, color = brightblue))
 #' @param legend_pos DEFAULT = 't' for top; Other legend positions are 'b', 'tr', 'l', 'r', and 'n' for none.
 #' @param legend_text_size DEFAULT = 16
@@ -48,7 +48,7 @@
 #' )
 #'
 #' chart <- ms_grouped_y2()
-#' print(chart, preview = T)
+#' print(chart, preview = TRUE)
 
 ms_grouped_y2 <-  function(
   data = frequencies,
@@ -57,12 +57,12 @@ ms_grouped_y2 <-  function(
   group_var = 'group_var',
   axis_text_size = 14,
   axis_title_size = 18,
-  axis_x_display = T,
+  axis_x_display = TRUE,
   axis_x_label = '',
   axis_x_position = c('nextTo', 'high', 'low', 'none'),
   axis_x_rotate = 0,
   axis_x_rotate_title = 0,
-  axis_y_display = T,
+  axis_y_display = TRUE,
   axis_y_label = '',
   axis_y_min = 0,
   axis_y_max = NULL,
@@ -74,7 +74,7 @@ ms_grouped_y2 <-  function(
   grouping = 'standard',
   label_color = color_settings_grouped,
   label_position = c('outEnd', 'inEnd', 'ctr', 'inBase'),
-  label_show_values = T,
+  label_show_values = TRUE,
   label_text = text_settings_grouped,
   overlapping = -25,
   legend_pos = c('t', 'n', 'b', 'tr', 'l', 'r'),
@@ -86,7 +86,7 @@ ms_grouped_y2 <-  function(
 
 ### Check for special symbols
   freqs_list <- split(data, seq(nrow(data))) # turn data frame into a list
-  symbols_sum <- purrr::map_df(freqs_list, ~str_detect(.x, "<|&")) %>% # test if any cells contain special symbols
+  symbols_sum <- purrr::map_df(freqs_list, ~stringr::str_detect(.x, "<|&")) %>% # test if any cells contain special symbols
     dplyr::mutate_all(~as.numeric(.)) %>% # convert table into numerics
     sum() # sum all cells to count the number of special symbols
   if(symbols_sum > 0){