+++ R/ms_single_y2.R
@@ -10,7 +10,7 @@
 #' @param axis_text_size DEFAULT = 14; Font size for variable levels and percentages.
 #' @param axis_title_size DEFAULT = 18; Font size for axis_x_label and axis_y_label.
 #' @param axis_x_position DEFAULT = 'nextTo'; Other options include "high", "low", "none". Change to "low" if dealing with negative numbers
-#' @param axis_x_display,axis_y_display DEFAULT = T
+#' @param axis_x_display,axis_y_display DEFAULT = TRUE
 #' @param axis_x_label,axis_y_label DEFAULT = ''; Title for the x_axis and y_axis
 #' @param axis_y_min DEFAULT = 0 to show full data without skewing perspective, but can be adjusted.
 #' @param axis_y_max DEFAULT = NULL
@@ -22,7 +22,7 @@
 #' @param grouping DEFAULT = 'standard'; grouping for a barchart, a linechart or an area chart. must be one of "percentStacked", "clustered", "standard" or "stacked".
 #' @param label_text DEFAULT = text_settings; A list of text settings for the percent labels. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: text_settings <- list("result" = fp_text(font.size = 10.5, color = bluepurple, font.family = 'Arial'))
 #' @param label_color DEFAULT = color_settings; A list of color settings for the bars. This affects font size and color. Specified outside of the function. If a list of one, no need to specify values. Otherwise, they must exactly match the group_var levels. Example: color_settings <- list(bluepurple)
-#' @param label_show_values DEFAULT = T; TRUE or FALSE. Show percent labels for each value.
+#' @param label_show_values DEFAULT = TRUE; TRUE or FALSE. Show percent labels for each value.
 #' @param label_position DEFAULT = 'outEnd'; Other options include c('outEnd', 'inEnd', 'ctr', 'inBase')
 #' @param legend_pos DEFAULT = 'n' for none. Other legend positions are 'b', 'tr', 'l', 'r', 't'.
 #' @param legend_text_size DEFAULT = 16
@@ -36,13 +36,13 @@
 #' @examples
 #' frequencies <- mtcars %>%
 #'   y2clerk::freqs(carb) %>%
-#'   orderlabel::order_label(inherent_order_label = T)
+#'   orderlabel::order_label(inherent_order_label = TRUE)
 #'
 #' color_settings <- list('blue')
 #' text_settings<- list('result' = fp_text(font.size = 20))
 #'
 #' chart <- ms_single_y2()
-#' print(chart, preview = T)
+#' print(chart, preview = TRUE)
 
 
 
@@ -52,12 +52,12 @@ ms_single_y2 <-  function(
   y_var = 'result',
   axis_text_size = 14,
   axis_title_size = 18,
-  axis_x_display = T,
+  axis_x_display = TRUE,
   axis_x_label = '',
   axis_x_position = c('nextTo', 'high', 'low', 'none'),
   axis_x_rotate = 0,
   axis_x_rotate_title = 0,
-  axis_y_display = T,
+  axis_y_display = TRUE,
   axis_y_label = '',
   axis_y_min = 0,
   axis_y_max = NULL,
@@ -70,7 +70,7 @@ ms_single_y2 <-  function(
   group_var = NULL,
   label_color = color_settings,
   label_position = c('outEnd', 'inEnd', 'ctr', 'inBase'),
-  label_show_values = T,
+  label_show_values = TRUE,
   label_text = text_settings,
   legend_pos = c('n', 't', 'b', 'tr', 'l', 'r'),
   legend_text_size = 16,