+++ tests/testthat/test_gg_single_y2.R
@@ -1,4 +1,10 @@
-#### order_label ####
+# Setup -------------------------------------------------------------------
+library(dplyr)
+library(y2clerk)
+library(testthat)
+library(y2artisan)
+
+#### gg_single_y2 ####
 context("gg_single_y2")